Uniform buffers beginning

+++ PBRVulkan/RayTracer/src/Tracer/Scene.cpp
@@ -22,7 +22,7 @@ namespace Tracer
 			{ { -0.5f, 0.5f, 0.0f }, { 0.0f, 1.0f, 0.0f }, { 0.5f, 0.5f } }
 		};
 
-		const std::vector<uint16_t> indices = {
+		const std::vector<uint32_t> indices = {
 			0, 1, 2, 2, 3, 0
 		};
